test: fix openapi.json cross-references by making them HTTP from FS
The relative paths were not being evaluated correctly and while
investigating that I realized that we should not use relative
paths at all.

Initially it made sense for rapid prototyping because this way
multiple openapi.json specs can be updated in the same commit, but
the price we pay is that the packages impicitly depend on each other
direclty via the file-system at runtime which is a big no-no.

By using the HTTP URLs instead, it is a little less flexible when
it comes to changes, but that isn't always a bad thing when it
comes to code stability.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>